Split filter and attribution-rule tests out of test_reporting
